+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Deprecated
 
 - `HandlerWithDetails` function. Use `WithDetails` instead.
+- `keyvals.ToMap` function. Use [emperror.dev/errors/utils/keyval.ToMap](https://godoc.org/emperror.dev/errors/utils/keyval/emperror#ToMap) instead.
 
 
 ## [0.21.0] - 2019-07-17